Add dry-run and keep-zip flags to upload script

+++ plugin_upload.py
@@ -23,10 +23,12 @@ ENDPOINT = '/plugins/RPC2/'
 VERBOSE = False
 
 ZIP_FILE_NAME = 'plugin.zip'
+PLUGIN_NAME = "Transition-QGIS"
 
 
 def zip_dir():
     """Zip plugin source code and place it inside a 'Transition-QGIS' folder"""
+    print(f"Zipping plugin source code into {ZIP_FILE_NAME}")
     with zipfile.ZipFile(ZIP_FILE_NAME, 'w', zipfile.ZIP_DEFLATED) as zipf:
         for root, dirs, files in os.walk(".", topdown=True):
             # Exclude certain directories from being zipped
@@ -36,13 +38,14 @@ def zip_dir():
                     # Calculate the relative path of the file to the root directory
                     relative_path = os.path.relpath(os.path.join(root, file), start=".")
                     # Construct the new path inside the 'Transition-QGIS' folder
-                    new_path = os.path.join("Transition-QGIS", relative_path)
+                    new_path = os.path.join(PLUGIN_NAME, relative_path)
                     # Write the file to the zip with the new path
                     zipf.write(os.path.join(root, file), arcname=new_path)
 
 
 def delete_zip():
     """Deletes the created zip file"""
+    print(f"Deleting zip file {ZIP_FILE_NAME}")
     if os.path.exists(ZIP_FILE_NAME):
         os.remove(ZIP_FILE_NAME)
         print(f"{ZIP_FILE_NAME} has been deleted.")
@@ -56,6 +59,9 @@ def main(parameters):
     :param parameters: Command line parameters.
     :param arguments: Command line arguments.
     """
+    # Zip plugin source code
+    zip_dir()
+
     address = "{protocol}://{username}:{password}@{server}:{port}{endpoint}".format(
         protocol=PROTOCOL,
         username=parameters.username,
@@ -63,20 +69,19 @@ def main(parameters):
         server=parameters.server,
         port=parameters.port,
         endpoint=ENDPOINT)
-    print("Connecting to: %s" % hide_password(address))
+    print("Connecting to QGIS plugin server at %s" % hide_password(address))
 
     server = xmlrpc.client.ServerProxy(address, verbose=VERBOSE)
     
-    # Zip plugin source code
-    zip_dir()
-
     try:
         # Upload plugin to QGIS
-        with open(ZIP_FILE_NAME, 'rb') as handle:
-            plugin_id, version_id = server.plugin.upload(
-                xmlrpc.client.Binary(handle.read()))
-        print("Plugin ID: %s" % plugin_id)
-        print("Version ID: %s" % version_id)
+        print("Uploading plugin to QGIS plugin server")
+        if not options.dry_run:
+            with open(ZIP_FILE_NAME, 'rb') as handle:
+                plugin_id, version_id = server.plugin.upload(
+                    xmlrpc.client.Binary(handle.read()))
+            print("Plugin ID: %s" % plugin_id)
+            print("Version ID: %s" % version_id)
     except xmlrpc.client.ProtocolError as err:
         print("A protocol error occurred")
         print("URL: %s" % hide_password(err.url, 0))
@@ -88,8 +93,9 @@ def main(parameters):
         print("Fault code: %d" % err.faultCode)
         print("Fault string: %s" % err.faultString)
 
-    # Delete zip file
-    delete_zip()
+    if not options.keep_zip:
+        # Delete zip file
+        delete_zip()
 
 
 def hide_password(url, start=6):
@@ -123,6 +129,13 @@ if __name__ == "__main__":
     parser.add_option(
         "-s", "--server", dest="server",
         help="Specify server name", metavar="plugins.qgis.org")
+    parser.add_option(
+        "--dry-run", dest="dry_run", action="store_true",
+        help="Perform a dry run without making any changes")
+    parser.add_option(
+        "--keep-zip", dest="keep_zip", action="store_true",
+        help="Keep the zip file after the script finishes instead of deleteing it")
+    
     options, args = parser.parse_args()
 
     if not options.server:
@@ -142,4 +155,7 @@ if __name__ == "__main__":
     if not options.password:
         # interactive mode
         options.password = getpass.getpass()
+    if options.dry_run:
+        print("Performing a dry run. No real upload will be made.")
+
     main(options)
